requirements.txt
# Facial Stress Indicator - Requirements

# Deep Learning
torch>=1.12.0
torchvision>=0.13.0

# Computer Vision
opencv-python>=4.6.0
Pillow>=9.0.0

# Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars
tqdm>=4.64.0

# Optional: Web UI
# streamlit>=1.12.0
# gradio>=3.0.0



